+++ src/example.py
@@ -1,38 +1,23 @@
 # Testing code
 
+import sys
 import uuid
 import datetime
 from io import BytesIO, StringIO
-from dotenv import load_dotenv
+import time
+
 from python_ms_core import Core
 from python_ms_core.core.queue.providers import azure_queue_config
 from python_ms_core.core.queue.models.queue_message import QueueMessage
-
-load_dotenv()
 
 Core.initialize()
 print('Hello')
 
 topic = 'gtfs-flex-upload'
-subscription = 'uploadprocessor'
+subscription = 'upload-validation-processor-test'
 some_other_sub = 'usdufs'
 
 topic_config = azure_queue_config.AzureQueueConfig()
-
-topicObject1 = Core.get_topic(topic_name=topic)
-topicObject2 = Core.get_topic(topic_name='tipic')
-
-print()
-
-queue_message = QueueMessage.data_from({
-    'message': str(uuid.uuid4().hex),
-    'data': {'a': 1}
-})
-
-topicObject1.publish(data=queue_message)
-
-msg = topicObject1.subscribe(subscription=subscription)
-print(f'Received message: {msg}')
 
 azure_client = Core.get_storage_client()
 
@@ -61,3 +46,26 @@ logger.record_metric(name='test', value='test')
 
 logger = Core.get_logger(provider='Local')
 logger.record_metric(name='test', value='test')
+
+
+def on_connect_callback(instance):
+    print('Connected with result code {}'.format(instance.get_messages()))
+    time.sleep(30)
+    sys.exit()
+
+
+topicObject1 = Core.get_topic(topic_name=topic, callback=on_connect_callback)
+topicObject2 = Core.get_topic(topic_name='tipic')
+try:
+    topicObject2.subscribe(subscription=subscription)
+except Exception as e:
+    print(e)
+
+queue_message = QueueMessage.data_from({
+    'message': str(uuid.uuid4().hex),
+    'data': {'a': 1}
+})
+
+topicObject1.publish(data=queue_message)
+
+topicObject1.subscribe(subscription=subscription)