+++ src/python_ms_core/core/resource_errors/errors.py
@@ -1,12 +1,16 @@
 class ServiceError(Exception):
+    status = 400
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         if args:
-            self.message = args[0]
+            self.message = args[0]['message']
+            self.status_code = args[0]['status_code'] if args[0]['status_code'] != 0 else self.status
+            self.description = args[0]['description']
         else:
             self.message = None
-
-    status = 400
+            self.status_code = self.status
+            self.description = None
 
 
 class BadRequestError(ServiceError):