+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+### 0.0.2
+- Added filtered exception handling
+- Removed unused packages
+- Fixed Topic subscription
+
 ### 0.0.1
 - Introduces Topic and subscription methods in Core.
 - Added methods